adding some copy function to the scene object

+++ src/render/render_image.c
@@ -6,7 +6,7 @@
 /*   By: tatahere <tatahere@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 12:13:34 by tatahere          #+#    #+#             */
-/*   Updated: 2025/03/06 16:51:16 by tatahere         ###   ########.fr       */
+/*   Updated: 2025/03/12 12:53:37 by tatahere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,9 +44,9 @@ void		render_image(mlx_image_t *img, t_scene *scene)
 	{
 		while (i < WIN_WIDTH)
 		{
-			ray.z = (j * 100.0 / (double) WIN_HEIGHT - 50.0) *  -1.0;
+			ray.z = (j * 100.0 / WIN_HEIGHT - 50.0) *  -1.0;
 			//ray.z = (double)((double)(j * 100) / (double)WIN_HEIGHT - (double)50) * (double) -1;
-			ray.y = i * 100.0 / (double) WIN_WIDTH - 50.0;
+			ray.y = i * 100.0 / WIN_WIDTH - 50.0;
 			//ray.y = (double)(i * 100) / (double)WIN_WIDTH - (double)50;
 			ray.x = tan(deg_to_rad(90.0 - (scene->camera.fov_rad / 2.0))) * 50.0;
 			ray = vec3_normalize(ray);
